project/dev-requirements.txt
pytest==3.3.1
pytest-django==3.1.2
model-mommy==1.5.0
